requirements.txt
channels==4.0.0
channels-redis==4.0.0
daphne==4.0.0
django-phonenumber-field==7.1.0
django-cors-headers
djangorestframework-simplejwt
dj_database_url==2.1.0
phonenumbers==8.13.17
Pillow==9.3.0
psycopg2-binary==2.9.5
pytest-asyncio==0.20.1
pytest-django==4.5.2
python-dotenv==1.0.0
